requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
pytest>=7.0
numpy>=1.24
matplotlib>=3.7
